parseFloat instead of parseInt
in runComparison function, I used parseFloat to make sure to compare even the floats in the ratings.

+++ index.js
@@ -84,9 +84,9 @@ const runComparison = () =>{
     // select every two corresponding article and compare them
     leftSideStats.forEach( (leftStat , index) =>{
         const rightStat = rightSideStats[index];
-
-        const leftSideValue = parseInt(leftStat.dataset.value);
-        const rightSideValue = parseInt(rightStat.dataset.value);
+        
+        const leftSideValue = parseFloat(leftStat.dataset.value);
+        const rightSideValue = parseFloat(rightStat.dataset.value);
 
         if(rightSideValue > leftSideValue){
             leftStat.classList.remove('is-primary');
